feat: add fake user repository and change floats to decimals

+++ user_service/src/domain/users.rs
@@ -1,16 +1,18 @@
-#[derive(Debug)]
+use rust_decimal::Decimal;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Gender {
     Male,
     Female,
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub struct User {
     pub username: String,
     pub first_name: String,
     pub last_name: String,
     pub age: u8,
-    pub weight: f32,
-    pub height: f32,
+    pub weight: Decimal,
+    pub height: Decimal,
     pub gender: Gender,
 }
